Don't ask "Stop learning?" if you choose to not forget a move
If you choose "YES" to forget a move but then go back, the question will still show up normally and if you then choose to not stop learning, it will send you back to the move selection instead of going through the dialogue again. Also, pressing B on that particular question will now be interpreted as "Yes" (Stop learning).